Creates ComapeoSettings folder. Moves the items into it based on UI.
