+++ src/frontend/screens/ComapeoSettings/CoordinateFormat.tsx
@@ -2,15 +2,15 @@ import * as React from 'react';
 import {ScrollView} from 'react-native';
 import {defineMessages, useIntl} from 'react-intl';
 
-import {formatCoords} from '../../../lib/coordinateFormat';
-import {SelectOne} from '../../../sharedComponents/SelectOne';
-import {type CoordinateFormat as CoordinateFormatType} from '../../../lib/coordinateFormat';
-import type {NativeNavigationComponent} from '../../../sharedTypes/navigation';
-import {useLastKnownLocation} from '../../../hooks/useLastSavedLocation';
+import {formatCoords} from '../../lib/coordinateFormat';
+import {SelectOne} from '../../sharedComponents/SelectOne';
+import {type CoordinateFormat as CoordinateFormatType} from '../../lib/coordinateFormat';
+import type {NativeNavigationComponent} from '../../sharedTypes/navigation';
+import {useLastKnownLocation} from '../../hooks/useLastSavedLocation';
 import {
   useCoordinateFormatActions,
   useCoordinateFormat,
-} from '../../../contexts/CoordinateFormatStoreContext';
+} from '../../contexts/CoordinateFormatStoreContext';
 
 const m = defineMessages({
   title: {
