+++ src/frontend/screens/ComapeoSettings/DataAndPrivacy/SettingsPrivacyPolicy.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import {ScrollView, StyleSheet} from 'react-native';
 import {defineMessages} from 'react-intl';
-import {PrivacyPolicy} from '../PrivacyPolicy';
+import {PrivacyPolicy} from '../../PrivacyPolicy';
 
 const m = defineMessages({
   navTitle: {
